Move edit/add log forms into logs page directory, add plus icon to add log button
