+++ react-app/src/components/LogsPage/EditLogFormModal/EditLogForm.js
@@ -2,26 +2,33 @@ import React, { useState, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useLocation, useHistory } from "react-router-dom";
 
-import { addLog } from "../../store/logs";
+import { editLog } from "../../../store/logs";
 
-const AddLogForm = ({ showModal, exercise_id, exerciseName}) => {
+const unitConverter = {
+    "Reps": 1,
+    "Weight": 2,
+    "Time": 3,
+    "Distance": 4
+}
+
+const EditLogForm = ({ showModal, exerciseName, eId, eDate, eUnit, eUnitCount, eComment}) => {
 	const dispatch = useDispatch();
 	const location = useLocation();
 	const history = useHistory();
 
 	const sessionUser = useSelector((state) => state.session.user);
 
-    let today = new Date();
-    const offset = today.getTimezoneOffset();
-    today = new Date(today.getTime() - (offset*60*1000));
-    today = today.toISOString().split('T')[0];
+    // let today = new Date();
+    // const offset = today.getTimezoneOffset();
+    // today = new Date(today.getTime() - (offset*60*1000));
+    // today = today.toISOString().split('T')[0];
 
 
 	const [errors, setErrors] = useState({});
-	const [date, setDate] = useState(today);
-    const [unit_id, setUnitId] = useState(1);
-	const [unit_count, setUnitCount] = useState(1);
-	const [comment, setComment] = useState("");
+	const [date, setDate] = useState(eDate);
+    const [unit_id, setUnitId] = useState(unitConverter[eUnit]);
+	const [unit_count, setUnitCount] = useState(eUnitCount);
+	const [comment, setComment] = useState(eComment);
 
 
     useEffect(() => {
@@ -45,9 +52,9 @@ const AddLogForm = ({ showModal, exercise_id, exerciseName}) => {
 
 	const handleSubmit = async (e) => {
 		e.preventDefault();
-		const data = await dispatch(addLog(
+		const data = await dispatch(editLog(
+            eId,
 			date,
-			exercise_id,
 			unit_id,
 			unit_count,
 			comment
@@ -90,7 +97,7 @@ const AddLogForm = ({ showModal, exercise_id, exerciseName}) => {
 
 	return (
 		<form className="form-container log-form" onSubmit={handleSubmit}>
-            <h3 className="form-heading">Log {exerciseName}</h3>
+            <h3 className="form-heading">Edit Log {exerciseName}</h3>
 			<div className="form-group">
 				<label className="form-label" htmlFor="date">
 					Date
@@ -168,4 +175,4 @@ const AddLogForm = ({ showModal, exercise_id, exerciseName}) => {
 	);
 };
 
-export default AddLogForm;
+export default EditLogForm;
